requirements.txt
python-telegram-bot[job-queue]==20.7
requests
beautifulsoup4
python-dotenv
psycopg2-binary
tzdata
